+++ jim-interactive.c
@@ -87,7 +87,7 @@ int Jim_InteractivePrompt(Jim_Interp *interp)
 
     home = getenv("HOME");
     if (home && isatty(STDIN_FILENO)) {
-        int history_len = strlen(home) + sizeof("/.jim_history");
+        int history_len = (int)(strlen(home) + sizeof("/.jim_history"));
         history_file = Jim_Alloc(history_len);
         snprintf(history_file, history_len, "%s/.jim_history", home);
         Jim_HistoryLoad(history_file);
