+++ jim-win32compat.c
@@ -46,7 +46,7 @@ int gettimeofday(struct timeval *tv, void *unused)
     struct _timeb tb;
 
     _ftime(&tb);
-    tv->tv_sec = tb.time;
+    tv->tv_sec = (long)tb.time;
     tv->tv_usec = tb.millitm * 1000;
 
     return 0;
@@ -75,7 +75,7 @@ DIR *opendir(const char *name)
             strchr("/\\", name[base_length - 1]) ? "*" : "/*";
 
         if ((dir = (DIR *) Jim_Alloc(sizeof *dir)) != 0 &&
-            (dir->name = (char *)Jim_Alloc(base_length + strlen(all) + 1)) != 0) {
+            (dir->name = (char *)Jim_Alloc((int)(base_length + strlen(all) + 1))) != 0) {
             strcat(strcpy(dir->name, name), all);
 
             if ((dir->handle = (long)_findfirst(dir->name, &dir->info)) != -1)
